+++ inc/entity/monster_nest.h
@@ -8,7 +8,7 @@ class MonsterNest : public Structure, public MonsterSpawnNode
 public:
     MonsterNest(int ix_, int iy_);
 
-    virtual float getShotVariance() { return 0.3f * width; }
+    virtual float getShotVariance() const { return 0.3f * width; }
 
 #ifdef DEBUG
     virtual void debugPrint() const;