+++ inc/entity/structure.h
@@ -20,7 +20,7 @@ public:
     virtual void receiveBombDamage(const ShotData& shot, double damage);
     virtual double calculateIncomingDamage(double damage, double crit);
 
-    virtual bool canBeTargeted() { return !isIndestructible; }
-    virtual float getShotVariance() { return 0.125f * width; }
-    virtual bool canLeech() { return false; }
+    virtual bool canBeTargeted() const { return !isIndestructible; }
+    virtual float getShotVariance() const { return 0.125f * width; }
+    virtual bool canLeech() const { return false; }
 };