+++ src/entity/beacon.cpp
@@ -17,10 +17,12 @@ Beacon::Beacon(int ix_, int iy_, BEACON_TYPE beaconType_)
     if (g_game.game == GC_LABYRINTH)
     {
         range = 100.0f / 33.0f;
+        nextShotTimer = rand() % 100 + 200;
     }
     else
     {
         range = 50.0f / 17.0f + 60.0f / 17.0f * (rand() / float(RAND_MAX));
+        nextShotTimer = rand() % 240 + 180;
     }
 }
 
@@ -28,6 +30,45 @@ bool Beacon::tick(IngameMap& map, int frames)
 {
     if (isKilled)
         return true;
+
+    if (!isPassive())
+    {
+        nextShotTimer -= frames;
+        if (nextShotTimer <= 0)
+        {
+            switch (beaconType)
+            {
+                case BEACON_HEAL:
+                    activateHeal(map);
+                    break;
+                case BEACON_SPEED:
+                    activateSpeed(map);
+                    break;
+                case BEACON_CLEANSE:
+                    activateCleanse(map);
+                    break;
+                case BEACON_INVULNERABILITY: // case BEACON_SHIELD:
+                    if (g_game.game == GC_LABYRINTH)
+                        activateInvulnerability(map);
+                    else
+                        activateShield(map);
+                    break;
+                case BEACON_MANA_BIND: // case BEACON_DISCHARGE:
+                    if (g_game.game == GC_LABYRINTH)
+                        activateManaBind(map);
+                    else
+                        activateDischarge(map);
+                    break;
+                case BEACON_OMNI:
+                    activateOmni(map);
+                    break;
+            }
+            if (g_game.game == GC_LABYRINTH)
+                nextShotTimer = rand() % 100 + 200;
+            else
+                nextShotTimer = rand() % 240 + 180;
+        }
+    }
 
     return false;
 }
@@ -53,6 +94,91 @@ void Beacon::receiveBombDamage(const ShotData& shot, double damage)
         damage *= 0.25;
 
     Structure::receiveBombDamage(shot, damage);
+}
+
+void Beacon::activateHeal(IngameMap& map)
+{
+    std::vector<Targetable*> targets = map.enemyController.getBeaconTargetsWithinRangeSq(y, x, range * range, false);
+
+    for (Targetable* t : targets)
+    {
+        Monster* pMonster = reinterpret_cast<Monster*>(t);
+
+        double heal;
+        if (g_game.game == GC_LABYRINTH)
+            heal = 0.3 * (1.0 - (110.0 / 66.0) * sqrt((t->x - x) * (t->x - x) + (t->y - y) * (t->y - y)));
+        else
+            heal = 0.05;
+
+        pMonster->receiveBeaconHeal(t->hpMax * heal);
+    }
+}
+
+void Beacon::activateSpeed(IngameMap& map)
+{
+    std::vector<Targetable*> targets = map.enemyController.getBeaconTargetsWithinRangeSq(y, x, range * range, true);
+    for (Targetable* t : targets)
+    {
+        Monster* pMonster = reinterpret_cast<Monster*>(t);
+        pMonster->receiveBeaconHaste();
+    }
+}
+
+void Beacon::activateCleanse(IngameMap& map)
+{
+    std::vector<Targetable*> targets = map.enemyController.getBeaconTargetsWithinRangeSq(y, x, range * range, false);
+    for (Targetable* t : targets)
+    {
+        Monster* pMonster = reinterpret_cast<Monster*>(t);
+        pMonster->receiveBeaconCleanse();
+    }
+}
+
+void Beacon::activateInvulnerability(IngameMap& map)
+{
+    std::vector<Targetable*> targets = map.enemyController.getBeaconTargetsWithinRangeSq(y, x, range * range, false);
+    for (Targetable* t : targets)
+    {
+        Monster* pMonster = reinterpret_cast<Monster*>(t);
+        pMonster->receiveBeaconInvulnerability();
+    }
+}
+
+void Beacon::activateManaBind(IngameMap& map)
+{
+    std::vector<Targetable*> targets = map.enemyController.getBeaconTargetsWithinRangeSq(y, x, range * range, true);
+    for (Targetable* t : targets)
+    {
+        Monster* pMonster = reinterpret_cast<Monster*>(t);
+        pMonster->receiveBeaconManaBind();
+    }
+}
+
+void Beacon::activateShield(IngameMap& map)
+{
+    std::vector<Targetable*> targets = map.enemyController.getBeaconTargetsWithinRangeSq(y, x, range * range, false);
+    for (Targetable* t : targets)
+    {
+        Monster* pMonster = reinterpret_cast<Monster*>(t);
+        pMonster->receiveBeaconShield();
+    }
+}
+
+void Beacon::activateDischarge(IngameMap& map)
+{
+    map.buildingController.dischargeGems(x, y, range * range);
+}
+
+void Beacon::activateOmni(IngameMap& map)
+{
+    std::vector<Targetable*> targets = map.enemyController.getBeaconTargetsWithinRangeSq(y, x, range * range, false);
+    for (Targetable* t : targets)
+    {
+        Monster* pMonster = reinterpret_cast<Monster*>(t);
+        pMonster->receiveBeaconHeal(t->hpMax * 0.05);
+        pMonster->receiveBeaconHaste();
+        pMonster->receiveBeaconCleanse();
+    }
 }
 
 #ifdef DEBUG